Share common code between JS backends
