+++ effekt/shared/src/main/scala/effekt/generator/js/JavaScript.scala
@@ -46,7 +46,7 @@ class JavaScript extends Compiler[String] {
     case input @ CoreTransformed(source, tree, mod, core) =>
       val mainSymbol = Context.checkMain(mod)
       val mainFile = path(mod)
-      val doc = pretty(TransformerDS.compile(input, mainSymbol).commonjs)
+      val doc = pretty(TransformerDirect.compile(input, mainSymbol).commonjs)
       (Map(mainFile -> doc.layout), mainFile)
   }
 
@@ -54,7 +54,7 @@ class JavaScript extends Compiler[String] {
   // -----------------------------------
   lazy val Separate:  Phase[Source, (CoreTransformed, Module)] =
     allToCore(Core) andThen all(core.LambdaLifting) map { in =>
-        (in.main, TransformerDS.compileSeparate(in))
+        (in.main, TransformerDirect.compileSeparate(in))
     }
 
   private def pretty(stmts: List[js.Stmt]): Document =
